requirements.txt
channels==2.1.3
Django==2.1.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
coverage==4.5.1
django-nose==1.4.6

